change package namespace from nl.dtls to org.fairdatateam
