+++ src/main/java/org/fairdatateam/rdf/migration/database/RdfMigrationRepository.java
@@ -20,44 +20,17 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package nl.dtls.rdf.migration.entity;
+package org.fairdatateam.rdf.migration.database;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import org.fairdatateam.rdf.migration.entity.RdfMigration;
+import org.springframework.data.mongodb.repository.MongoRepository;
 
 /**
- * An annotation serves for marking classes which should contain a concrete migration. Its
- * properties contain a meta-information about the migration.
+ * A Mongo repository for manipulating with {@link RdfMigration}
  *
  * @author Vojtech Knaisl (vknaisl)
  * @since 1.0.0
  */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface RdfMigrationAnnotation {
-
-    /**
-     * A number of the migration in your application ({@link nl.dtls.rdf.migration.entity.RdfMigration#number})
-     *
-     * @return A getter for the value
-     */
-    int number();
-
-    /**
-     * A name of the migration ({@link nl.dtls.rdf.migration.entity.RdfMigration#name})
-     *
-     * @return A getter for the value
-     */
-    String name();
-
-    /**
-     * A quick description of the purpose of the migration
-     * ({@link nl.dtls.rdf.migration.entity.RdfMigration#description})
-     *
-     * @return A getter for the value
-     */
-    String description();
+public interface RdfMigrationRepository extends MongoRepository<RdfMigration, String> {
 
 }
